pyproject.toml
[tool.poetry]
name = "HypeSquad"
version = "0.1.0"
description = ""
authors = ["SeoYoungDeok <0duck_seo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.11.4"
torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-linux_x86_64.whl"}
opencv-python = "^4.7.0.72"
numpy = "^1.24.3"
matplotlib = "^3.7.1"
pyyaml = "^6.0"
ipykernel = "^6.23.1"
albumentations = "^1.3.0"
wandb = "^0.15.3"
hydra-core = "^1.3.2"
hydra-zen = "^0.10.1"
timm = "^0.9.2"
tqdm = "^4.65.0"
rich = "^13.4.1"
torchmetrics = "1.0.0"
lightning = "^2.0.2"
openmim = "^0.3.7"
segmentation-models-pytorch = "^0.3.3"
pandas = "^2.0.2"
transformers = "^4.30.2"
optuna = "^3.2.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
